services/voice_cloner_service/requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
audioread==3.1.0
brotli==1.2.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.6
chatterbox-tts==0.1.6
click==8.3.1
colorama==0.4.6
conformer==0.3.2
decorator==5.2.1
Deprecated==1.3.1
diffusers==0.29.0
distlib==0.4.0
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.135.2
ffmpy==1.0.0
filelock==3.25.2
fsspec==2026.2.0
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
identify==2.6.18
idna==3.11
importlib_metadata==9.0.0
jaconv==0.5.0
Jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
nodeenv==1.10.0
numba==0.64.0
numpy==1.25.2
omegaconf==2.3.0
onnx==1.20.1
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.9.4
pooch==1.9.0
pre_commit==4.5.1
protobuf==7.34.1
pycparser==3.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pykakasi==2.3.0
pyloudnorm==0.2.0
python-dateutil==2.9.0.post0
python-discovery==1.2.0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.33.0
resemble-perth==1.0.1
rich==14.3.3
ruff==0.15.7
s3tokenizer==0.3.0
safehttpx==0.1.7
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
spacy_pkuseg==1.0.1
srsly==2.5.3
starlette==0.52.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.20.3
tomlkit==0.13.3
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
tqdm==4.67.3
transformers==4.46.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
virtualenv==21.2.0
websockets==15.0.1
wrapt==2.1.2
zipp==3.23.0
